pyproject.toml
[project]
name = "desafio-catskillet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.4",
    "django-redis>=6.0.0",
    "django-stubs[compatible-mypy]>=5.2.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "redis>=6.2.0",
    "ruff>=0.12.5",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"
